requirements.txt
asyncpg
aiofiles
aiohttp
python-dotenv
openai
numpy
scikit-learn
chromadb
pyignite
faiss-cpu
torch
networkx
python-json-logger
